GUI diagnostics tool and reccomendations
Diagnoses button is finished. A user can now click on the diagnoses button to check the actual status and attempt to identify a problem himself. A reccomendation is also given based on the actual situation. Some commands are also shown to the user so that it promotes self resolution of most common  issues.