Added logger, updated logic and added param
Added logger to all functions. Altered load_on_start, so that it takes a dictionary of params, enabling it to pass ip information that was previously gathered, to avoid unnecessary IP calls this delaying the response time of the GUI. This change was also made in update_labels_server_list.