Added concurrent.futures and updated other methods
Concurrent futures is used so that a message dialog is displayed to the user while the GUI is starting up. This will wait for a response from load_on_start, if it responds True then it will go ahead and load and content, otherwise it will inform the user that there might be connectivity problems. Other methods were updated, since openvpn_connect will return a tuple (message, servers_list). By doing this, it speeds up the conneciton process but not making multiple file reads.